refactor: extracrt keyboard shorcut actions into EditorKeyboardShortcutsActions outside of QuillRawEditorState
