+++ lib/src/editor/raw_editor/keyboard_shortcuts/editor_keyboard_shortcuts.dart
@@ -13,7 +13,7 @@ import '../../../document/nodes/line.dart';
 import '../../widgets/keyboard_listener.dart';
 import '../config/events/character_shortcuts_events.dart';
 import '../config/events/space_shortcut_events.dart';
-import 'default_single_activator_actions.dart';
+import 'default_single_activator_intents.dart';
 
 @internal
 class EditorKeyboardShortcuts extends StatelessWidget {
